Wonderful😁
This project aims to assess the risk of a recession in the U.S. economy between 1975 and 2021 based on seven economic indicators, including the House Price Index, Stock Price Index, Consumer Price Index, GDP, Unemployment Rate, Disposable Income, and Mortgage Rate. The study first visualizes US economic trends through time-series graphs and verifies economic theories and models using US economic data. The second part of the study compares the results of two approaches used to determine the risk of a recession: the yield curve and a machine learning algorithm. The machine learning algorithm, which uses the Random Forest algorithm in the “caret” and “randomForest” packages in R, was far from predicting a recession effectively. The conclusion drawn from the results and analysis is that although the machine learning algorithm can provide valuable insights, it is not an effective method for predicting a recession. But open a new opportunity for recession risk assessment. The study highlights the importance of combining multiple techniques and considering various economic indicators when assessing the risk of a recession.